lib: store a future in a static
Using a nightly channel, we can make use of the
`type_alias_impl_trait` feature that let's us
refer to anonymous compiler-generated types. In
turn this allows us to declare a static into
which we can store an instance of such a type.

We do this to store a future type into a static.
In addition we use a mutex for safe sync interior
mutability. The future is stored behind an Option,
which allows us to initialize it at compile-time
with a valid bit-pattern (None), alternatively,
we could use `core::mem::MaybeUninit`.

We can't use TAIT directly on a async function,
and must rather wrap it in a function where we
can specify an appropriate impl return type.