lib: poll the stored future in step()
Provide a const fn that returns a noop `RawWaker`,
that is then used when polling the stored future
during the `step()` call.

The signature of `step()` was now changed so that
it returns a bool, namely `true` to signal that
the state machine has reached its terminal state.

The FFI function can be safely called afterwards
as well, but it won't result in any computation
with the state machine, until we reinitialize it
with a call to `init`.